refactor: photoModal을 common -> specific 디렉토리로 이동
