+++ frontend/src/components/specific/photoModal/PhotoModal.styles.ts
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
 // biome-ignore format: 의도적으로 한 줄로 유지
-import { DownloadIcon as SaveIcon, TrashCanIcon } from '../../../../@assets/icons';
-import { theme } from '../../../../styles/theme';
-import { hexToRgba } from '../../../../utils/hexToRgba';
+import { DownloadIcon as SaveIcon, TrashCanIcon } from '../../../@assets/icons';
+import { theme } from '../../../styles/theme';
+import { hexToRgba } from '../../../utils/hexToRgba';
 
 export const Wrapper = styled.div`
   position: relative;
