+++ frontend/src/pages/guest/imageUploadPage/ImageUploadPage.tsx
@@ -9,10 +9,10 @@ import FloatingActionButton from '../../../components/@common/buttons/floatingAc
 import FloatingIconButton from '../../../components/@common/buttons/floatingIconButton/FloatingIconButton';
 import HighlightText from '../../../components/@common/highlightText/HighlightText';
 import GuestImageGrid from '../../../components/@common/imageLayout/imageGrid/guestImageGrid/GuestImageGrid';
-import GuestPhotoModal from '../../../components/@common/modal/photoModal/GuestPhotoModal';
 import UserBadge from '../../../components/@common/userBadge/UserBadge';
 import GuestSpaceHeader from '../../../components/layout/header/spaceHeader/guestSpaceHeader/GuestSpaceHeader';
 import LoadingLayout from '../../../components/layout/loadingLayout/LoadingLayout';
+import GuestPhotoModal from '../../../components/specific/photoModal/GuestPhotoModal';
 import UploadBox from '../../../components/specific/uploadBox/UploadBox';
 import { ROUTES } from '../../../constants/routes';
 import { useOverlay } from '../../../contexts/OverlayProvider';
